backend/requirements.txt
django
djangorestframework
gtts
requests
openai
django-cors-headers
scikit-learn
pandas
numpy
joblib
gunicorn
psycopg2-binary
whitenoise
python-decouple
dj-database-url
